Set up as cmake subproject.
