+++ CMakeLists.txt
@@ -12,5 +12,5 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-add_subdirectory(include)
+add_subdirectory(src)
 add_subdirectory(proto)
